requirements.txt
monai
torchio
git+https://github.com/MIC-DKFZ/MedNeXt.git#egg=mednext